requirements.txt
setuptools
python-shout==0.2.6
python-twitter==3.5
mutagen==1.43.0
pyliblo==0.10.0
pycurl==7.43.0.2
pyyaml==3.13
mysqlclient==1.4.6
xmltodict==0.12.0
-e .
